# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Tracker API",
    "description": "\n    ## Flock Tracker API\n\n    A comprehensive API for tracking farm animals, events, and locations.\n\n    ### Features:\n    * **Animal Management**: Track sheep, chickens, and bees with lineage information\n    * **Event Tracking**: Record health events, treatments, breeding, and more\n    * **Location Management**: Manage paddocks and locations with capacity tracking\n    * **Search & Filter**: Powerful search and filtering capabilities\n\n    ### Animal Types:\n    * **Sheep**: Track lineage, health events, and location with gender tracking (ewes/rams)\n    * **Chickens**: Monitor flock health and egg production\n    * **Hives**: Track hive health and honey production\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/receipts/upload": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Upload Receipt",
        "description": "Upload a receipt image or PDF for OCR processing",
        "operationId": "upload_receipt_api_receipts_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_receipt_api_receipts_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/receipts/{receipt_id}/process": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Process Receipt",
        "description": "Process a receipt with OCR and extract structured data (async with Celery)",
        "operationId": "process_receipt_api_receipts__receipt_id__process_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/task/{task_id}": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Get Task Status",
        "description": "Check the status of an OCR processing task",
        "operationId": "get_task_status_api_receipts_task__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Get Receipts",
        "description": "Get all receipts",
        "operationId": "get_receipts_api_receipts__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceiptResponse"
                  },
                  "title": "Response Get Receipts Api Receipts  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/{receipt_id}": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Get Receipt",
        "description": "Get a specific receipt",
        "operationId": "get_receipt_api_receipts__receipt_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "receipts"
        ],
        "summary": "Delete Receipt",
        "description": "Delete a receipt and its file",
        "operationId": "delete_receipt_api_receipts__receipt_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_receipt_api_receipts_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_receipt_api_receipts_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReceiptResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "extracted_data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Data"
          },
          "expense_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "file_type",
          "id",
          "file_path",
          "created_at",
          "updated_at"
        ],
        "title": "ReceiptResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}